Unit counts are 64-bit; they were rolling over on a real plan
⚠️ A Palatine Keepstar wants 1.35 billion Pyerite and rather more Tritanium. Every quantity
in the production calculator was an int, so the totals wrapped: the grid showed
-2,147,483,639 Tritanium at a cost of -8,246,337,174 ISK, and "Missing" pinned at exactly
2,147,483,647 — int.MaxValue, which is the tell.

Widened end to end rather than patched at the display: ProductionQueueEntry.Quantity, the
job's QuantityNeeded / Runs / QuantityProduced / Leftover, the material line's EffQtyPerRun /
TotalQty / Available / Missing, the raw and leftover quantities, the intermediate and final
product counts, both raw pools, ExpandItem's qty, the chain-material tuple, and
IndustryMe.JobMaterialTotal. The compiler found each site once the models moved.

Left as int on purpose: type ids, ME levels, and the SDE's own per-run recipe quantity and
portion size, none of which count units of a plan.

⚠️ Inventory levels stay int — that is the column type in the rules table — so the two places
that push a chain into a level clamp rather than cast. An absurd plan produces a capped level
instead of a negative one.

ISK is decimal and volume is double, so neither needed widening.

Also on this pass: a Raw Material Volume line on the Summary tab under Net Cost, summed from
the raw materials only — intermediates are built on site, not hauled in — and the Jobs tab
material columns widened from 70px to fit twelve digits, with padding so adjacent numbers
cannot run together.

+++ Services/IndustryMe.cs
@@ -46,9 +46,9 @@ public static class IndustryMe
     /// a job eats. This was two identical private copies — one here, one in BuildCostService
     /// — each carrying a comment telling the reader to keep it in step with the other.</para>
     /// </summary>
-    public static int JobMaterialTotal(int baseQty, double factor, int runs)
+    public static long JobMaterialTotal(int baseQty, double factor, long runs)
     {
         double total = Math.Round(baseQty * (double)runs * factor, 2);
-        return Math.Max(runs, (int)Math.Ceiling(total));
+        return Math.Max(runs, (long)Math.Ceiling(total));
     }
 }
